requirements.txt
django==4.0.3
mysqlclient==2.1.0
django-mapbox-location-field==2.0.0